Произведено разбиение на модули
