+++ package.json
@@ -5,13 +5,21 @@
   "description": "Учебный проект «Менеджер задач» от HTML Academy",
   "scripts": {
     "eslint": "eslint src/**/*.js",
-    "test": "npm run eslint"
+    "test": "npm run eslint",
+    "build": "webpack",
+    "start": "webpack-dev-server --open"
   },
   "devDependencies": {
     "eslint": "6.1.0",
-    "eslint-config-htmlacademy": "0.5.2"
+    "eslint-config-htmlacademy": "0.5.2",
+    "webpack": "^4.39.2",
+    "webpack-cli": "3.3.7"
   },
   "engines": {
     "node": "10"
+  },
+  "dependencies": {
+    "@webpack-cli/init": "^0.2.2",
+    "webpack-dev-server": "^3.8.0"
   }
 }
