+++ public/index.html
@@ -46,6 +46,6 @@
 
       <!-- Контент -->
     </main>
-    <script src="../src/main.js"></script>
+    <script src="bundle.js"></script>
   </body>
 </html>
