+++ src/components/board.js
@@ -1,0 +1,7 @@
+const getBoardElement = () => {
+  return `
+  <section class="board container"></section>
+  `;
+};
+
+export {getBoardElement};
