+++ src/components/button.js
@@ -1,0 +1,7 @@
+const getLoadMoreButtonElement = () => {
+  return `
+  <button class="load-more" type="button">load more</button>
+  `;
+};
+
+export {getLoadMoreButtonElement};
