+++ src/components/card-container.js
@@ -1,0 +1,7 @@
+const getCardContainerElement = () => {
+  return `
+  <div class="board__tasks"></div>
+  `;
+};
+
+export {getCardContainerElement};
